change value from float to double add sl4j

+++ src/main/java/com/mynt/deliveryparcel/db/domain/ParcelDetails.java
@@ -26,6 +26,6 @@ public class ParcelDetails {
     private Float condition;
 
     @Column(name = "cost")
-    private Float cost;
+    private Double cost;
 
 }
