+++ src/main/java/com/mynt/deliveryparcel/dto/request/ParcelDetailsRequest.java
@@ -19,22 +19,22 @@ public class ParcelDetailsRequest {
     @Min(value = 1, message = "Weight should be at least 1 kg.")
     @Max(value = 50, message = "Weight should not exceed 50 kgs.")
     @Schema(description = "weight of the parcel", example = "50")
-    private Float weight;
+    private Double weight;
 
     @NotNull(message = "Please input valid height.")
     @Min(value = 1, message = "Weight should be at least 1 cm.")
     @Schema(description = "Height of the parcel", example = "20")
-    private Float height;
+    private Double height;
 
     @NotNull(message = "Please input valid width.")
     @Min(value = 1, message = "Weight should be at least 1 cm.")
     @Schema(description = "Width of the parcel", example = "30")
-    private Float width;
+    private Double width;
 
     @NotNull(message = "Please input valid length.")
     @Min(value = 1, message = "Weight should be at least 1 cm.")
     @Schema(description = "Length of the parcel", example = "5")
-    private Float length;
+    private Double length;
 
     @Schema(description = "Voucher Code", example = "MYNT")
     private String voucherCode;
